Refactor fragment cache methods
Removed extra calls to constantize and DRY'd the code.
